RD-4899: Extracted consts and utils into separate files
