+++ app/widgets/common/deployments/DeploymentActionsMenu.tsx
@@ -5,58 +5,10 @@ import type { Workflow } from '../executeWorkflow';
 import { Menu, Popup, PopupMenu } from '../../../components/basic';
 import StageUtils from '../../../utils/stageUtils';
 import type { Label } from '../labels/types';
+import { menuItems } from './DeploymentActionsMenu.consts';
+import { isMenuItemAvailable } from './DeploymentActionsMenu.utils';
 
 const translate = StageUtils.getT('widgets.common.deployments.actionsMenu');
-
-export const actions = Object.freeze({
-    delete: 'delete',
-    forceDelete: 'forceDelete',
-    install: 'install',
-    manageLabels: 'manageLabels',
-    setSite: 'setSite',
-    uninstall: 'uninstall',
-    update: 'update',
-    deployOn: 'deployOn'
-});
-
-const permissions = {
-    executeWorkflowPermission: 'execution_start',
-    deploymentDelete: 'deployment_delete',
-    deploymentUpdateCreate: 'deployment_update_create',
-    deploymentSetSite: 'deployment_set_site',
-    deploymentCreate: 'deployment_create'
-};
-
-type MenuItem = { name: string; icon: string; permission: string };
-const menuItems: MenuItem[] = [
-    { name: actions.install, icon: 'play', permission: permissions.executeWorkflowPermission },
-    { name: actions.update, icon: 'edit', permission: permissions.deploymentUpdateCreate },
-    { name: actions.deployOn, icon: 'rocket', permission: permissions.deploymentCreate },
-    { name: actions.setSite, icon: 'building', permission: permissions.deploymentSetSite },
-    { name: actions.manageLabels, icon: 'tags', permission: permissions.deploymentCreate },
-    { name: actions.uninstall, icon: 'recycle', permission: permissions.executeWorkflowPermission },
-    { name: actions.delete, icon: 'trash alternate', permission: permissions.deploymentDelete },
-    { name: actions.forceDelete, icon: 'trash', permission: permissions.deploymentDelete }
-];
-
-function isDeployOnFunctionalityAvailable(deploymentLabels: Label[]) {
-    return deploymentLabels.some(deploymentLabel => {
-        return deploymentLabel.key === 'csys-obj-type' && deploymentLabel.value === 'environment';
-    });
-}
-
-function isMenuItemAvailable(item: MenuItem, workflows: Workflow[], deploymentLabels: Label[]) {
-    if (item.permission === permissions.executeWorkflowPermission) {
-        const workflow = workflows?.find(w => w.name === item.name);
-        return !!workflow?.is_available;
-    }
-
-    if (item.name === actions.deployOn) {
-        return isDeployOnFunctionalityAvailable(deploymentLabels);
-    }
-
-    return true;
-}
 
 interface DeploymentActionsMenuProps {
     onActionClick: (actionName: string) => void;
